Separate hooks based on contract and update imports
