+++ src/hooks/radioDAONFT/index.ts
@@ -1,5 +1,2 @@
 export { useTokenURI } from "./useTokenURI";
-export { useGetAllNFTsForSale } from "./useGetAllNFTsForSale";
 export { useGetMyNFTs } from "./useGetMyNFTs";
-export { useGetMarketplaceFee } from "./useGetMarketplaceFee";
-export { useApproveAndSellNFT } from "./useApproveAndSellNFT";
